Implemented progress handling with BaseViewController and BaseViewModel
Added BaseViewController class, which is a subclass of UIViewController and conforms to LoadingView protocol. This class handles progress indication using a CustomProgress animator.

Also, added BaseViewModel class, which provides isLoading observable to indicate whether progress is ongoing or not. The observeLoading method in BaseViewController observes this isLoading property and updates UI accordingly.